+++ packages/pos/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [10.0.1](https://github.com/stone-payments/pos-mamba-sdk/compare/v10.0.0...v10.0.1) (2023-07-12)
+
+**Note:** Version bump only for package @mamba/pos
+
+
+
+
+
+## [10.0.0](https://github.com/stone-payments/pos-mamba-sdk/compare/v9.7.0...v10.0.0) (2023-07-10)
+
+
+### Features
+
+* 🎸 [wip] org drivers ([1a82f0d](https://github.com/stone-payments/pos-mamba-sdk/commit/1a82f0d33f88254c9af4a8b86725dfc8703746ab))
+* 🎸 implements integration with organizations module ([0d7db60](https://github.com/stone-payments/pos-mamba-sdk/commit/0d7db600fa5d17b3ff4c66fa2e3076ec3cb069b8))
+* 🎸 implements synchronous getOrganizationFile ([a73f45c](https://github.com/stone-payments/pos-mamba-sdk/commit/a73f45c9db02fba8cf817730458d159e532ddfa8))
+
+
+
 ## [9.7.0](https://github.com/stone-payments/pos-mamba-sdk/compare/v9.6.0...v9.7.0) (2023-06-09)
 
 **Note:** Version bump only for package @mamba/pos