+++ packages/webpack/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.0.0](https://github.com/stone-payments/pos-mamba-sdk/compare/v9.7.0...v10.0.0) (2023-07-10)
+
+
+### Features
+
+* 🎸 [wip] org drivers ([1a82f0d](https://github.com/stone-payments/pos-mamba-sdk/commit/1a82f0d33f88254c9af4a8b86725dfc8703746ab))
+* 🎸 add favicon to simulator env ([a50963e](https://github.com/stone-payments/pos-mamba-sdk/commit/a50963ec381444c33ee67dde6a9472fb2ebf190f))
+* 🎸 implements integration with organizations module ([0d7db60](https://github.com/stone-payments/pos-mamba-sdk/commit/0d7db600fa5d17b3ff4c66fa2e3076ec3cb069b8))
+
+
+
 ## [9.7.0](https://github.com/stone-payments/pos-mamba-sdk/compare/v9.6.0...v9.7.0) (2023-06-09)